+++ dotfiles/.config/quickshell/shell.qml
@@ -1,12 +1,25 @@
 import Quickshell
+import Quickshell.Io
 import "WelcomeApp"
 import "PowerApp"
 import "SidebarApp"
 import "CalendarApp"
+import "WallpaperApp"
+import "CustomTheme"
 
 ShellRoot {
+    // Test IPC tools: qs ipc show
+
+    IpcHandler {
+        target: "theme-manager" 
+        function reload(): void {
+            Theme.reloadTheme()
+        }
+    }
+
     WelcomeWindow {}
     PowerWindow {}
     SidebarWindow {}
     CalendarWindow {}
+    WallpaperWindow {}
 }