+++ setup/post-arch.sh
@@ -1,7 +1,5 @@
 #!/usr/bin/env bash
 
-SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
-
 # --------------------------------------------------------------
 # Oh My Posh
 # --------------------------------------------------------------
@@ -12,45 +10,24 @@ curl -s https://ohmyposh.dev/install.sh | bash -s -- -d ~/.local/bin
 # --------------------------------------------------------------
 
 bash <(curl -s https://raw.githubusercontent.com/mylinuxforwork/ml4w-dotfiles-settings/main/setup.sh)
-rm $HOME/.local/share/ml4w-dotfiles-settings/quickshell/shared/Theme.qml  
-ln -sf $HOME/.config/quickshell/shared/Theme.qml $HOME/.local/share/ml4w-dotfiles-settings/quickshell/shared/Theme.qml
-
-# --------------------------------------------------------------
-# Temporary Installation of waypaper-git
-# --------------------------------------------------------------
-
-yay --noconfirm -Rns waypaper
-yay --noconfirm -S waypaper-git
-
-# --------------------------------------------------------------
-# Prebuilt Packages
-# --------------------------------------------------------------
-
-source $SCRIPT_DIR/_prebuilt.sh
 
 # --------------------------------------------------------------
 # Cursors
 # --------------------------------------------------------------
 
-source $SCRIPT_DIR/_cursors.sh
+source $repo_path/setup/_cursors.sh
 
 # --------------------------------------------------------------
 # Fonts
 # --------------------------------------------------------------
 
-source $SCRIPT_DIR/_fonts.sh
+source $repo_path/setup/_fonts.sh
 
 # --------------------------------------------------------------
 # Icons
 # --------------------------------------------------------------
 
-source $SCRIPT_DIR/_icons.sh
-
-# --------------------------------------------------------------
-# Migrate
-# --------------------------------------------------------------
-
-source $SCRIPT_DIR/migrate.sh
+source $repo_path/setup/_icons.sh
 
 # --------------------------------------------------------------
 # Create XDG Directories