+++ setup/post-fedora.sh
@@ -1,7 +1,5 @@
 #!/usr/bin/env bash
 
-SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
-
 # --------------------------------------------------------------
 # Oh My Posh
 # --------------------------------------------------------------
@@ -13,18 +11,31 @@ curl -s https://ohmyposh.dev/install.sh | bash -s -- -d ~/.local/bin
 # --------------------------------------------------------------
 
 bash <(curl -s https://raw.githubusercontent.com/mylinuxforwork/ml4w-dotfiles-settings/main/setup.sh)
-rm $HOME/.local/share/ml4w-dotfiles-settings/quickshell/shared/Theme.qml  
-ln -sf $HOME/.config/quickshell/shared/Theme.qml $HOME/.local/share/ml4w-dotfiles-settings/quickshell/shared/Theme.qml
 
 # --------------------------------------------------------------
-# Prebuild Packages
+# Cargo
 # --------------------------------------------------------------
 
-source $SCRIPT_DIR/_prebuilt.sh
+TARGET_VERSION="4.0.0"
 
-echo "Installing eza v0.23.0"
-# https://github.com/eza-community/eza/releases
-sudo cp $SCRIPT_DIR/packages/eza /usr/bin
+force_install_matugen() {
+    info "Running: cargo install matugen --force"
+    cargo install matugen --force
+}
+
+if ! command -v matugen &> /dev/null; then
+    echo "'matugen' is not currently installed."
+    force_install_matugen
+else
+    CURRENT_VERSION=$(matugen --version | grep -Eo '[0-9]+\.[0-9]+\.[0-9]+' | head -n 1)
+    LOWEST_VERSION=$(printf "%s\n%s" "$TARGET_VERSION" "$CURRENT_VERSION" | sort -V | head -n1)
+    if [ "$LOWEST_VERSION" = "$CURRENT_VERSION" ] && [ "$CURRENT_VERSION" != "$TARGET_VERSION" ]; then
+        info "Current version ($CURRENT_VERSION) is lower than $TARGET_VERSION. Updating..."
+        force_install_matugen
+    else
+        info "matugen is already up to date! (Current version: $CURRENT_VERSION)"
+    fi
+fi
 
 # --------------------------------------------------------------
 # Pip
@@ -32,49 +43,30 @@ sudo cp $SCRIPT_DIR/packages/eza /usr/bin
 
 echo ":: Installing packages with pip"
 sudo pip install pywalfox
-sudo pip install screeninfo
-sudo pip install waypaper
-
-# --------------------------------------------------------------
-# TTY Clock
-# --------------------------------------------------------------
-
-# git clone https://github.com/xorg62/tty-clock
-# cd tty-clock
-# sudo dnf install ncurses ncurses-devel -y
-# make
-# chmod +x tty-clock
-# sudo mv tty-clock /usr/local/bin/tty-clock
 
 # --------------------------------------------------------------
 # Grimblast
 # --------------------------------------------------------------
 
-sudo cp $SCRIPT_DIR/scripts/grimblast /usr/bin
+sudo cp $repo_path/setup/scripts/grimblast /usr/bin
 
 # --------------------------------------------------------------
 # Cursors
 # --------------------------------------------------------------
 
-source $SCRIPT_DIR/_cursors.sh
+source $repo_path/setup/_cursors.sh
 
 # --------------------------------------------------------------
 # Fonts
 # --------------------------------------------------------------
 
-source $SCRIPT_DIR/_fonts.sh
+source $repo_path/setup/_fonts.sh
 
 # --------------------------------------------------------------
 # Icons
 # --------------------------------------------------------------
 
-source $SCRIPT_DIR/_icons.sh
-
-# --------------------------------------------------------------
-# Migrate
-# --------------------------------------------------------------
-
-source $SCRIPT_DIR/migrate.sh
+source $repo_path/setup/_icons.sh
 
 # --------------------------------------------------------------
 # Create XDG Directories