+++ setup/preflight-fedora.sh
@@ -8,8 +8,9 @@ sudo dnf copr enable --assumeyes wef/cliphist
 sudo dnf copr enable --assumeyes tofik/nwg-shell
 sudo dnf copr enable --assumeyes che/nerd-fonts
 sudo dnf copr enable --assumeyes erikreider/SwayNotificationCenter
+sudo dnf copr enable --assumeyes mineiro/hyprland
+# Quickshell
 sudo dnf copr enable --assumeyes errornointernet/quickshell
-sudo dnf copr enable --assumeyes mineiro/hyprland 
 
 # --------------------------------------------------------------
 # Uninstall swww if exists. To be replaced with awww in the next steps