+++ setup/preflight-opensuse.sh
@@ -10,16 +10,17 @@ curl -s https://ohmyposh.dev/install.sh | bash -s -- -d ~/.local/bin
 # Prebuild Packages
 # --------------------------------------------------------------
 
-source $SCRIPT_DIR/_prebuilt.sh
+source $repo_path/setup/_prebuilt.sh
 
 # --------------------------------------------------------------
 # Repositories
 # --------------------------------------------------------------
 
-sudo zypper addrepo https://download.opensuse.org/repositories/home:AvengeMedia:danklinux/openSUSE_Tumbleweed/home:AvengeMedia:danklinux.repo
 sudo zypper addrepo https://download.opensuse.org/tumbleweed/repo/oss/ factory-oss
 sudo zypper addrepo https://download.opensuse.org/repositories/X11:Wayland/openSUSE_Tumbleweed/X11:Wayland.repo
 sudo zypper addrepo https://download.opensuse.org/repositories/X11:fonts/openSUSE_Factory/X11:fonts.repo
+sudo zypper addrepo https://download.opensuse.org/repositories/home:/Alxhr0/openSUSE_Tumbleweed/ home_Alxhr0
+sudo zypper refresh
 
 # --------------------------------------------------------------
 # Uninstall swww if exists. To be replaced with awww in the next steps